# components/manager/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Fetcher Manager component",
        "title": "Fetcher Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:4006",
    "basePath": "/",
    "paths": {
...
        "/v1/management/connections/validate-schema": {
            "post": {
                "description": "Validate that tables and fields referenced in the request exist in the configured datasources. Returns 200 when validation passes, 422 when validation fails with detailed error information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connections"
                ],
                "summary": "Validate schema",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "X-Organization-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Schema validation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.SchemaValidationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Validation successful - all tables and fields exist",
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.SchemaValidationResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or missing headers",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Validation failed - schema errors found (missing tables, fields, or unreachable datasources)",
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.SchemaValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/v1/management/connections/{id}/schema": {
            "get": {
                "description": "Get the database schema (tables and fields) for a connection.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connections"
                ],
                "summary": "Get connection schema",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "X-Organization-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.ConnectionSchemaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/management/connections/{id}/test": {
            "post": {
                "description": "Test the configured connection by establishing and closing a connection.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connections"
                ],
                "summary": "Test connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "X-Organization-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connection test result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_LerianStudio_fetcher_pkg_model.ConnectionSchemaResponse": {
            "type": "object",
            "properties": {
                "configName": {
                    "type": "string"
                },
                "databaseName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.TableDetails"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_LerianStudio_fetcher_pkg_model.SchemaValidationError": {
            "type": "object",
            "properties": {
                "dataSourceId": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.SchemaValidationErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.SchemaValidationError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.SchemaValidationRequest": {
            "description": "Request body for schema validation containing mapped fields per datasource.",
            "type": "object",
            "required": [
                "mappedFields"
            ],
            "properties": {
                "mappedFields": {
                    "description": "MappedFields maps datasource config names to their tables and fields\nKey: configName (e.g., \"midaz_onboarding\")\nValue: map of table names to field names",
                    "type": "object",
                    "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.SchemaValidationResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_LerianStudio_fetcher_pkg_model.SchemaValidationError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" or \"failure\"",
                    "type": "string"
                }
            }
        },
        "github_com_LerianStudio_fetcher_pkg_model.TableDetails": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pkg.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                },
                "err": {},
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}